pyproject.toml
[tool.poetry]
name = "codeaireview"
version = "0.1.0"
description = ""
authors = ["ivanmishakivskiy <imishakivskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.6"
pydantic = "^2.10.3"
requests = "^2.32.3"
openai = "^1.57.4"
redis = "^5.2.1"
httpx = "^0.28.1"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.0"
pytest-asyncio = "^0.25.0"
pytest-mock = "^3.14.0"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]